GoogleAuth/Users/serializers.py
from rest_framework import serializers
from .models import *


class UsersSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id','email','first_name','last_name','phone_number','county','birthdate','city','gender','address']
